+++ astropy/stats/tests/test_funcs.py
@@ -16,9 +16,21 @@ except ImportError:
 else:
     HAS_SCIPY = True
 
+try:
+    import mpmath  # pylint: disable=W0611
+except ImportError:
+    HAS_MPMATH = False
+else:
+    HAS_MPMATH = True
+
 from ...tests.helper import pytest
 
 from .. import funcs
+
+# These are not part of __all__ because they are just the lower level versions
+# of poisson_upper_limit
+#from ..funcs import scipy_poisson_upper_limit, mpmath_poisson_upper_limit
+
 from ...utils.misc import NumpyRNGContext
 from ... import units as u
 
@@ -481,3 +493,69 @@ def test_poisson_conf_gehrels86(n):
         funcs.poisson_conf_interval(
             n, interval='frequentist-confidence')[1],
         rtol = 0.02)
+
+@pytest.mark.skipif('not HAS_SCIPY')
+def test_scipy_poisson_limit():
+    '''Test that the lower-level routine gives the snae number.
+
+    Test numbers are from table1 1, 3 in
+    Kraft, Burrows and Nousek in
+    `ApJ 374, 344 (1991) <http://adsabs.harvard.edu/abs/1991ApJ...374..344K>`_
+    '''
+    assert_allclose(funcs._scipy_kraft_burrows_nousek(5., 2.5, .99), (0, 10.67), rtol=1e-3)
+    conf = funcs.poisson_conf_interval([5., 6.], 'kraft-burrows-nousek',
+                                       background=[2.5, 2.],
+                                       conflevel=[.99, .9])
+    assert_allclose(conf[:, 0], (0, 10.67), rtol=1e-3)
+    assert_allclose(conf[:, 1], (0.81, 8.99), rtol=5e-3)
+
+
+@pytest.mark.skipif('not HAS_MPMATH')
+def test_mpmath_poisson_limit():
+    assert_allclose(funcs._mpmath_kraft_burrows_nousek(6., 2., .9), (0.81, 8.99), rtol=5e-3)
+    assert_allclose(funcs._mpmath_kraft_burrows_nousek(5., 2.5, .99), (0, 10.67), rtol=1e-3)
+
+@pytest.mark.skipif('not HAS_SCIPY')
+def test_poisson_conf_value_errors():
+    with pytest.raises(ValueError) as e:
+        funcs.poisson_conf_interval([5, 6], 'root-n', sigma=2)
+    assert 'Only sigma=1 supported' in str(e.value)
+
+    with pytest.raises(ValueError) as e:
+        funcs.poisson_conf_interval([5, 6], 'pearson', background=[2.5, 2.])
+    assert 'background not supported' in str(e.value)
+
+    with pytest.raises(ValueError) as e:
+        funcs.poisson_conf_interval([5, 6], 'sherpagehrels', conflevel=[2.5, 2.])
+    assert 'conflevel not supported' in str(e.value)
+
+    with pytest.raises(ValueError) as e:
+        funcs.poisson_conf_interval(1, 'foo')
+    assert 'Invalid method' in str(e.value)
+
+
+@pytest.mark.skipif('not HAS_SCIPY')
+def test_poisson_conf_kbn_value_errors():
+    with pytest.raises(ValueError) as e:
+        funcs.poisson_conf_interval(5., 'kraft-burrows-nousek',
+                                    background=2.5,
+                                    conflevel=99)
+    assert 'number between 0 and 1' in str(e.value)
+
+    with pytest.raises(ValueError) as e:
+        funcs.poisson_conf_interval(5., 'kraft-burrows-nousek',
+                                    background=2.5)
+    assert 'Set conflevel for method' in str(e.value)
+
+    with pytest.raises(ValueError) as e:
+        funcs.poisson_conf_interval(5., 'kraft-burrows-nousek',
+                                    background=-2.5,
+                                    conflevel=.99)
+    assert 'Background must be' in str(e.value)
+
+
+@pytest.mark.skipif('HAS_SCIPY or HAS_MPMATH')
+def test_poisson_limit_nodependencies():
+    with pytest.raises(ImportError) as e:
+        a = funcs.poisson_conf_interval(20.,  interval='kraft-burrows-nousek',
+                                        background=10., conflevel=.95)